XenAPI: Do not rely on XenAPI to get device names
The reason is in some version of xenserver, it assumes all HVM domains
are Windows and returns incorrect device name.
In all xenserver versions the value from XAPI is just a guess and the
guest kernel may decide to use different device name.
With the new method, there is no need of any hacks, such as for
maverick support, etc.

Closes-Bug: #1318544

DocImpact: Remove two config variables 'remap_vbd_dev' and
           'remap_vbd_dev_prefix' in nova/conf/xenserver.py
           The new implementation do not rely on XenAPI call and the
           two config variables are not used any more

Change-Id: Ia61d1c212a115168f8a22fc513b6baf2102a30d1